+++ modules/orca/output.tf
@@ -81,3 +81,9 @@ output "orca_sqs_status_update_queue_id" {
   description = "The URL ID of the status-update-queue SQS"
   value       = module.orca_sqs.orca_sqs_status_update_queue_id
 }
+## Secretsmanager Module Outputs (orca_secretsmanager)
+## =============================================================================
+output "orca_secretsmanager_arn" {
+  description = "The Amazon Resource Name (ARN) of the AWS secretsmanager"
+  value       = module.orca_secretsmanager.secretsmanager_arn
+}