+++ modules/orca/variables.tf
@@ -50,17 +50,6 @@ variable "workflow_config" {
 
 
 # OPTIONAL - Default variable value is set in ../variables.tf to keep default values centralized.
-variable "aws_profile" {
-  type        = string
-  description = "AWS profile used to deploy the terraform application."
-}
-
-
-variable "region" {
-  type        = string
-  description = "AWS region to deploy configuration to."
-}
-
 
 variable "tags" {
   type        = map(string)
@@ -70,25 +59,10 @@ variable "tags" {
 
 ## Variables unique to ORCA
 ## REQUIRED
-variable "db_admin_username" {
-  description = "Username for RDS database administrator authentication"
-  type        = string
-}
-variable "database_admin_name" {
-  description = "Name of RDS database administrator authentication"
-  type        = string
-}
+
+
 variable "db_admin_password" {
   description = "Password for RDS database administrator authentication"
-  type        = string
-}
-variable "database_user_name" {
-  description = "Name of RDS database user authentication"
-  type        = string
-}
-
-variable "db_user_username" {
-  description = "Username for RDS database user authentication"
   type        = string
 }
 variable "db_user_password" {
@@ -106,11 +80,10 @@ variable "orca_default_bucket" {
 }
 
 ## OPTIONAL - Default variable value is set in ../variables.tf to keep default values centralized.
-variable "database_port" {
-  type        = number
-  description = "Database port that PostgreSQL traffic will be allowed on."
+variable "db_admin_username" {
+  description = "Username for RDS database administrator authentication"
+  type        = string
 }
-
 
 variable "orca_ingest_lambda_memory_size" {
   type        = number
@@ -193,15 +166,13 @@ variable "status_update_queue_message_retention_time_seconds" {
   description = "The number of seconds status_update_queue SQS retains a message in seconds. Maximum value is 14 days."
 }
 
-## OPTIONAL (DO NOT CHANGE!) - Development use only
-## Default variable value is set in ../variables.tf to keep default values centralized.
-
-variable "database_name" {
+# variables needed by bootstrap lambda. Remove in ORCA-243
+variable "aws_profile" {
   type        = string
-  description = "Name of the ORCA database in PostgreSQL"
+  description = "AWS profile used to deploy the terraform application."
 }
 
-variable "orca_recovery_retrieval_type" {
+variable "region" {
   type        = string
-  description = "AWS glacier recovery type to use. One of Bulk, Standard, Express."
+  description = "AWS region to deploy configuration to."
 }