+++ modules/secretsmanager/main.tf
@@ -1,29 +1,10 @@
-# Local Variables
+## Local Variables
 locals {
   tags = merge(var.tags, { Deployment = var.prefix })
 }
 
 ## Resources
 
-## =============================================================================
-## NULL RESOURCES - 1x Use
-## =============================================================================
-
-## bootstrap - Bootstrap lambda that creates/modifies database objects on deploys
-## =============================================================================
-# https://registry.terraform.io/providers/hashicorp/null/latest/docs/resources/resource
-# https://www.terraform.io/docs/language/resources/provisioners/local-exec.html
-resource "null_resource" "bootstrap" {
-  # Determine what has to change to trigger this resource being created/re-created.
-  triggers = {
-    bootstrap_lambda_hash = var.db_deploy_source_code_hash
-  }
-
-  # Execute the db_deploy lambda 1 time if the resource is created/re-created.
-  provisioner "local-exec" {
-    command = "aws lambda invoke --function-name ${var.db_deploy_arn} 'db_deploy-response.out'"
-  }
-}
 ## =============================================================================
 ## SECRET MANAGER RESOURCES
 ## =============================================================================
@@ -41,15 +22,14 @@ resource "aws_secretsmanager_secret_version" "db_login" {
   secret_id  = aws_secretsmanager_secret.db_login.id
   depends_on = [aws_secretsmanager_secret.db_login]
   secret_string = jsonencode({
-    admin_username = var.db_admin_username
+    admin_username = "var.db_admin_username"
     admin_password = var.db_admin_password
-    admin_database = var.database_admin_name
-    user_username  = var.db_user_username
+    admin_database = "postgres"
+    user_username  = "orcauser"
     user_password  = var.db_user_password
-    user_database  = var.database_user_name
+    user_database  = "disaster_recovery"
     host           = var.db_host_endpoint
-    port           = var.database_port
-    tags           = local.tags
+    port           = "5432"
   })
 }
 ## KMS key policy
@@ -57,6 +37,7 @@ resource "aws_secretsmanager_secret_version" "db_login" {
 data "aws_iam_policy_document" "orca_kms_key_policy" {
   statement {
     sid = "orca_kms_policy"
+    # work with NGAP to restrict actions
     actions = [
       "kms:Encrypt",
       "kms:Decrypt",